template

+++ domain/item.go
@@ -3,7 +3,7 @@ package domain
 type Item struct {
 	Id         string
 	Title      string
-	Price      float32
+	Price      float64
 	CurrencyId string
 	Thumbnail  string
 	Permalink  string
